Refactor post-processing the linter result
Use `reposition_match` to fix or recompute the `col` the linter outputs.  By doing so we avoid accessing the buffer
which is a mutable, live object and thus a side-effect.  Note that assuming a tab-size of 4 is still idiosyncratic.

Simplify `split_match` to fill in the required `line` information for unspecific errors.